backend/requirements.txt
fastapi
uvicorn 
sqlalchemy 
pydantic[dotenv] 
pydantic[email] 
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.3.0
python-multipart
boto3==1.35.76
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
google-generativeai==0.8.3